requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-huggingface>=0.0.1
langchain-chroma>=0.0.1

# Vector database
chromadb>=0.4.18

# Embedding models
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0

# NLP processing
spacy>=3.6.0

# Knowledge graph
networkx>=3.1

# Document processing
unstructured>=0.10.0
pdfminer.six>=20221105
docx2txt>=0.8
beautifulsoup4>=4.12.0
markdownify>=0.11.0
python-magic>=0.4.27
PyMuPDF>=1.22.0

# API and Web
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
tqdm>=4.66.1
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: for progress bars in notebooks
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0